mcp-telegram/pyproject.toml
[project]
name = "mcp-telegram"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.100.0",
    "fastmcp>=2.1.2",
    "uvicorn>=0.22.0",
    "httpx>=0.28.0",
    "python-telegram-bot>=20.0.0",
    "httpx>=0.28.0"
]
